+++ mayan/apps/dynamic_search/tests/backends.py
@@ -3,7 +3,7 @@ from django.utils.module_loading import import_string
 from mayan.apps.storage.utils import TemporaryDirectory
 
 from ..classes import SearchBackend
-from ..backends.elastic_search import ElasticSearchBackend
+from ..backends.elasticsearch import ElasticSearchBackend
 from ..backends.whoosh import WhooshSearchBackend
 from ..literals import DEFAULT_SEARCH_BACKEND
 from ..settings import setting_backend_arguments
@@ -23,35 +23,38 @@ class TestSearchBackend(SearchBackend):
     It will also automatically insert the test object into the index
     before performing a call to the `search` method.
     """
-    _test_view = None
+    _test_backend_initialized = False
+    _test_backend_path = None
+    _test_class = None
 
     def __init__(self, *args, **kwargs):
-        if self._test_view:
-            backend_path = getattr(self._test_view, '_test_search_backend_path', DEFAULT_SEARCH_BACKEND)
+        if self._test_class:
+            backend_path = getattr(self._test_class, '_test_search_backend_path', DEFAULT_SEARCH_BACKEND)
         else:
             backend_path = DEFAULT_SEARCH_BACKEND
+            SearchBackend._disable()
 
-        self._backend_kwargs = setting_backend_arguments.value.copy()
-        self._backend_class = import_string(dotted_path=backend_path)
+        if self.__class__._test_backend_path != backend_path:
+            self.__class__._test_backend_path = backend_path
 
-        if issubclass(self._backend_class, WhooshSearchBackend):
-            if not hasattr(self.__class__, '_local_attribute_backend_temporary_directory'):
-                self.__class__._local_attribute_backend_temporary_directory = TemporaryDirectory()
+            self.__class__._backend_kwargs = setting_backend_arguments.value.copy()
+            self.__class__._backend_class = import_string(dotted_path=backend_path)
 
-            self._backend_kwargs['index_path'] = self._local_attribute_backend_temporary_directory.name
-            self._backend = self._backend_class(**self._backend_kwargs)
-        elif issubclass(self._backend_class, ElasticSearchBackend):
-            self._backend_kwargs['indices_namespace'] = 'test'
-            self._backend = self._backend_class(**self._backend_kwargs)
+            if issubclass(self._backend_class, WhooshSearchBackend):
+                if not hasattr(self.__class__, '_local_attribute_backend_temporary_directory'):
+                    self.__class__._local_attribute_backend_temporary_directory = TemporaryDirectory()
 
-            if not hasattr(self.__class__, '_local_attribute_backend_indices_cleared'):
-                self.__class__._local_attribute_backend_indices_cleared = True
-                self._backend.reset()
-        else:
-            self._backend = self._backend_class(**self._backend_kwargs)
+                self.__class__._backend_kwargs['index_path'] = self._local_attribute_backend_temporary_directory.name
+            elif issubclass(self._backend_class, ElasticSearchBackend):
+                self.__class__._backend_kwargs['indices_namespace'] = 'test'
 
-        if not self._test_view:
-            SearchBackend.uninitialize_class()
+            self.__class__._test_backend_initialized = False
+
+        self.__class__._backend = self._backend_class(**self.__class__._backend_kwargs)
+
+        if not self.__class__._test_backend_initialized:
+            self.__class__._test_backend_initialized = True
+            self._backend._initialize()
 
         super().__init__(*args, **kwargs)
 
@@ -59,12 +62,19 @@ class TestSearchBackend(SearchBackend):
         return self._backend._search(*args, **kwargs)
 
     def deindex_instance(self, *args, **kwargs):
-        if self._test_view:
+        if self._test_class:
             return self._backend.deindex_instance(*args, **kwargs)
 
     def index_instance(self, *args, **kwargs):
-        if self._test_view:
+        if self._test_class:
             return self._backend.index_instance(*args, **kwargs)
 
-    def reset(self):
-        return self._backend.reset()
+    def index_search_model(self, *args, **kwargs):
+        return self._backend.index_search_model(*args, **kwargs)
+
+    def reset(self, *args, **kwargs):
+        return self._backend.reset(*args, **kwargs)
+
+    def tear_down(self):
+        self.__class__._test_backend_path = None
+        return self._backend.tear_down()